+++ velox/connectors/hive/storage_adapters/s3fs/RegisterS3FileSystem.cpp
@@ -131,8 +131,6 @@ void registerS3Metrics() {
   DEFINE_METRIC(kMetricS3FailedUploads, velox::StatType::COUNT);
   DEFINE_METRIC(kMetricS3SuccessfulUploads, velox::StatType::COUNT);
   DEFINE_METRIC(kMetricS3MetadataCalls, velox::StatType::COUNT);
-  DEFINE_METRIC(kMetricS3ListStatusCalls, velox::StatType::COUNT);
-  DEFINE_METRIC(kMetricS3ListLocatedStatusCalls, velox::StatType::COUNT);
   DEFINE_METRIC(kMetricS3GetObjectErrors, velox::StatType::COUNT);
   DEFINE_METRIC(kMetricS3GetMetadataErrors, velox::StatType::COUNT);
   DEFINE_METRIC(kMetricS3GetObjectRetries, velox::StatType::COUNT);
