+++ velox/connectors/hive/storage_adapters/s3fs/S3Counters.h
@@ -1,5 +1,5 @@
 /*
-* Copyright (c) Facebook, Inc. and its affiliates.
+ * Copyright (c) Facebook, Inc. and its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,34 +13,27 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 #pragma once
 
 namespace facebook::velox::filesystems {
 
 constexpr std::string_view kMetricS3ActiveConnections{
-  "presto_cpp.hive.s3.active_connections"};
+    "presto_cpp.hive.s3.active_connections"};
 constexpr std::string_view kMetricS3StartedUploads{
-  "presto_cpp.hive.s3.started_uploads"};
+    "presto_cpp.hive.s3.started_uploads"};
 constexpr std::string_view kMetricS3SuccessfulUploads{
-  "presto_cpp.hive.s3.successful_uploads"};
+    "presto_cpp.hive.s3.successful_uploads"};
 constexpr std::string_view kMetricS3FailedUploads{
-  "presto_cpp.hive.s3.failed_uploads"};
+    "presto_cpp.hive.s3.failed_uploads"};
 constexpr std::string_view kMetricS3MetadataCalls{
-  "presto_cpp.hive.s3.metadata_calls"};
-constexpr std::string_view kMetricS3ListStatusCalls{
-  "presto_cpp.hive.s3.list_status_calls"};
-constexpr std::string_view kMetricS3ListLocatedStatusCalls{
-  "presto_cpp.hive.s3.list_located_status_calls"};
-constexpr std::string_view kMetricS3ListObjectsCalls{
-  "presto_cpp.hive.s3.list_objects_calls"};
+    "presto_cpp.hive.s3.metadata_calls"};
 constexpr std::string_view kMetricS3GetObjectErrors{
-  "presto_cpp.hive.s3.get_object_errors"};
+    "presto_cpp.hive.s3.get_object_errors"};
 constexpr std::string_view kMetricS3GetMetadataErrors{
-  "presto_cpp.hive.s3.get_metadata_errors"};
+    "presto_cpp.hive.s3.get_metadata_errors"};
 constexpr std::string_view kMetricS3GetObjectRetries{
-  "presto_cpp.hive.s3.get_object_retries"};
+    "presto_cpp.hive.s3.get_object_retries"};
 constexpr std::string_view kMetricS3GetMetadataRetries{
-  "presto_cpp.hive.s3.get_metadata_retries"};
+    "presto_cpp.hive.s3.get_metadata_retries"};
 
-} // namespace facebook::velox
+} // namespace facebook::velox::filesystems
