+++ velox/connectors/hive/storage_adapters/s3fs/tests/CMakeLists.txt
@@ -96,3 +96,12 @@ target_link_libraries(
   velox_exec
   GTest::gtest
   GTest::gtest_main)
+
+add_executable(velox_s3metrics_test S3FileSystemMetricsTest.cpp)
+add_test(velox_s3metrics_test velox_s3metrics_test)
+target_link_libraries(
+  velox_s3metrics_test
+  velox_s3fs
+  velox_exec_test_lib
+  GTest::gtest
+  GTest::gtest_main)
